move function `getDestroyPromise` into own file
